+++ src/ssh/HISTORY.md
@@ -4,6 +4,7 @@ Release History
 -----
 * Add support to Microsoft.ConnectedVMwarevSphere/virtualMachines Resource Type.
 * Correct the format of expected input for --resource-type parameter. From Resource Provider name (e.g. "Microsoft.HybridCompute") to Resource Type name (e.g. "Microsoft.HybridCompute/machines").
+* [bug fix] SSH Banners are printed before authentication.
 
 1.1.2
 -----